Check for existing files when creating app in current directory
Check for existing files when creating app in current directory. This is
smart and dynamic. It walks the sampleApp files and then walks the
current directory to see if any files are the same. If they are the same
it warns the user to ensure they don't override anything important.