Unit Testing (In Progress) (#132)
* Prompt Unit Tests 1

Unit tests to test user prompting with file generation.

* Unit Tests (In Progress)

Unit tests for many of the prompts.

* Rename of HTTPS port number variable

Change to the HTTPS Port Number variable because it was the same as the standard HTTP Port Number variable.

* Fix for Travis CI

Fix for Travis CI so the build passes.

* Added more unit tests

Added more unit tests for user prompts.

* Updated package-lock

Updated package-lock to fix travis ci.

* Fix for function timeout

Function timeout  fixed for build checks.

* Refactor input validation

This refactors input validation to store reusable validation functions into an object vs using anonymous functions like before.

* Unit tests for helper functions

Unit tests that test the helper functions.

* More refactoring

This commit creates one more helper function and refactors the view engine recursion to a do..while to pass unit tests.

* Revert to prior chooseViewEngine()

Revert back to prior chooseViewEngine() function and new unit test for it.

* Revert "Revert to prior chooseViewEngine()"

This reverts commit 1bcfff80420781e19fcc87cdb5165832e898e9fb.

* Removed failed unit test

* Updated chooseViewEngine() for unit tests.

This allows us to increment the prompt names when recursively calling the function so we can unit test this properly using yeoman-test.

* Updated package-lock for merge

* Refactored Unit Tests

Unit test refactoring to remove unnecessary tests while keeping coverage at 100%.

* Removed Extra Defaults

removed extra default prompts that aren't needed.

* Set AppVeyor node version to 10.1

This is because of the unit tests breaking when using yeoman-test with AppVeyor and Node 10.2 and higher. 10.1 is the latest stable version of Node that yeoman-test will work with AppVeyor.